Removed tfplan/plan.go
This file provided support for reading TF11 binary plan files, which is no longer supported by terraform-validator and which was incompatible with terraform-plugin-sdk/v2
